# back/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "MaxAirain Express API with Swagger",
    "description": "This is a simple CRUD API application made for MaxAirain Company, with Express and documented with Swagger",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users in system",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "description": "ID of user that we want to find",
          "type": "integer"
        }
      ],
      "get": {
        "tags": ["Users"],
        "summary": "Get user in system by his ID",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{adminId}/create-user": {
      "parameters": [
        {
          "name": "adminId",
          "in": "path",
          "required": true,
          "description": "ID of admin",
          "type": "integer"
        }
      ],
      "post": {
        "tags": ["Users"],
        "summary": "Created user in system",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login for user",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{adminId}/users/{userId}": {
      "parameters": [
        {
          "name": "adminId",
          "in": "path",
          "required": true,
          "description": "ID of user that we want to find",
          "type": "integer"
        },
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "description": "ID of the admin",
          "type": "integer"
        }
      ],
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user in system by his ID",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": ["first_name", "last_name", "role", "picture_face"],
      "properties": {
        "_id": {
          "type": "integer",
          "uniqueItems": true
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "post": {
          "type": "string"
        },
        "picture_face": {
          "type": "array"
        }
      }
    },
    "Users": {
      "type": "array",
      "$ref": "#/definitions/User"
    }
  }
}